requirements.txt
# AI Agent для локальных моделей
# Основные зависимости

# Для работы с Ollama API
requests>=2.31.0
aiohttp>=3.9.0

# Для работы с локальными моделями напрямую (опционально)
torch>=2.1.0
transformers>=4.35.0
accelerate>=0.25.0
bitsandbytes>=0.41.0

# Для веб-интерфейса
fastapi>=0.104.0
uvicorn>=0.24.0
jinja2>=3.1.2
python-multipart>=0.0.6

# Для работы с кодом
pygments>=2.16.0
tree-sitter>=0.20.4
tree-sitter-python>=0.20.4

# Утилиты
python-dotenv>=1.0.0
colorama>=0.4.6
rich>=13.7.0
prompt-toolkit>=3.0.43
pyyaml>=6.0.1

# Для оптимизации
sentencepiece>=0.1.99
protobuf>=4.25.0

